components/localization/requirements.txt
argparse
grpcio
matplotlib
numpy
protobuf
pyserial
requests
scipy
